Ajuste de float para double em conta devido a perda de dados.

+++ desafioNgBiling/src/main/java/com/ngbiling/desafioNgBiling/model/Conta.java
@@ -9,6 +9,6 @@ import lombok.Data;
 public class Conta {
     @Id
     private int numero_conta;
-    private float saldo;
+    private double saldo;
 
 }
